MainScreen: вынес connect в отдельный модуль
