+++ src/components/main-screen/index.js
@@ -1,0 +1,3 @@
+import Component from './main-screen.connect';
+
+export default Component;
